+++ src/range-navigator/range-navigator.ts
@@ -369,7 +369,7 @@ export class RangeNavigator extends Component<HTMLElement> {
     @Event()
     public resized: EmitType<IResizeRangeNavigatorEventArgs>;
 
-    /** 
+    /**
      * Triggers before the label rendering
      * @event
      */
@@ -596,7 +596,7 @@ export class RangeNavigator extends Component<HTMLElement> {
     public createSecondaryElement(): void {
         // For userInteraction
         if (this.tooltip.enable) {
-            let tooltipDiv: Element = document.createElement('div');
+            let tooltipDiv: Element = this.createElement('div');
             tooltipDiv.id = this.element.id + '_Secondary_Element';
             tooltipDiv.setAttribute('style', 'position: relative');
             this.element.appendChild(tooltipDiv);
@@ -614,7 +614,7 @@ export class RangeNavigator extends Component<HTMLElement> {
     }
 
     /**
-     * To Remove the SVG. 
+     * To Remove the SVG.
      * @return {boolean}
      * @private
      */
@@ -686,7 +686,7 @@ export class RangeNavigator extends Component<HTMLElement> {
     }
 
     /**
-     * Handles the widget resize. 
+     * Handles the widget resize.
      * @return {boolean}
      * @private
      */
@@ -721,7 +721,7 @@ export class RangeNavigator extends Component<HTMLElement> {
     }
 
     /**
-     * Handles the mouse move. 
+     * Handles the mouse move.
      * @return {boolean}
      * @private
      */
@@ -731,7 +731,7 @@ export class RangeNavigator extends Component<HTMLElement> {
         return false;
     }
     /**
-     * Handles the mouse leave. 
+     * Handles the mouse leave.
      * @return {boolean}
      * @private
      */
@@ -742,7 +742,7 @@ export class RangeNavigator extends Component<HTMLElement> {
         return false;
     }
     /**
-     * Handles the mouse click on range navigator. 
+     * Handles the mouse click on range navigator.
      * @return {boolean}
      * @private
      */
@@ -759,8 +759,8 @@ export class RangeNavigator extends Component<HTMLElement> {
     }
     /**
      * Handles the export method for range navigator control.
-     * @param type 
-     * @param fileName 
+     * @param type
+     * @param fileName
      */
     public export(
         type: ExportType, fileName: string, orientation?: PdfPageOrientation,
@@ -782,7 +782,7 @@ export class RangeNavigator extends Component<HTMLElement> {
     }
 
     /**
-     * Handles the mouse down on range navigator. 
+     * Handles the mouse down on range navigator.
      * @return {boolean}
      * @private
      */
@@ -792,7 +792,7 @@ export class RangeNavigator extends Component<HTMLElement> {
         return false;
     }
     /**
-     * Handles the mouse up. 
+     * Handles the mouse up.
      * @return {boolean}
      * @private
      */
@@ -834,8 +834,8 @@ export class RangeNavigator extends Component<HTMLElement> {
     }
     /**
      * OnProperty change method calling here
-     * @param newProp 
-     * @param oldProp 
+     * @param newProp
+     * @param oldProp
      */
     public onPropertyChanged(newProp: RangeNavigatorModel, oldProp: RangeNavigatorModel): void {
         let renderer: boolean = false;
@@ -904,7 +904,7 @@ export class RangeNavigator extends Component<HTMLElement> {
     /**
      * To provide the array of modules needed for control rendering
      * @return {ModuleDeclaration[]}
-     * @private 
+     * @private
      */
     public requiredModules(): ModuleDeclaration[] {
         let modules: ModuleDeclaration[] = [];